+++ generate.go
@@ -1,0 +1,4 @@
+package goapitosdk
+
+//go:generate go run ./cmd/apito-gen
+//go:generate go run github.com/Khan/genqlient